# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Recovery Velocity Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /communications:
    get:
      operationId: listCommunications
      tags: [communications]
      summary: List operational communications in the caller's authorized scope
      parameters:
        - in: query
          name: priority
          required: false
          schema: { type: string, enum: [emergency, urgent, operational, information] }
        - in: query
          name: channel
          required: false
          schema: { type: string, enum: [app, radio, sms] }
        - in: query
          name: audience
          required: false
          schema: { $ref: "#/components/schemas/CommunicationAudience" }
      responses:
        "200":
          description: Communications
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/CommunicationMessage" }
        "403": { description: Operational authority required }
    post:
      operationId: createCommunication
      tags: [communications]
      summary: Record a server-attributed operational communication
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CommunicationMessageInput" }
      responses:
        "201":
          description: Communication recorded
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CommunicationMessage" }
        "400": { description: Invalid input }
        "403": { description: Scope denied }
  /communications/{id}/acknowledgements:
    post:
      operationId: acknowledgeCommunication
      tags: [communications]
      summary: Acknowledge a communication once per authenticated operator
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Acknowledgement recorded or previously recorded
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CommunicationAcknowledgement" }
        "403": { description: Scope denied }
        "404": { description: Communication not found }
  /communications/public-alerts:
    get:
      operationId: listPublicAlerts
      tags: [communications]
      summary: List published unexpired public alerts without authentication
      parameters:
        - in: query
          name: countryCode
          required: true
          schema: { type: string }
        - in: query
          name: parishId
          required: false
          schema: { type: string }
      responses:
        "200":
          description: Published public alerts
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/PublicAlert" }
  /communications/public-alerts/pending:
    get:
      operationId: listPendingPublicAlerts
      tags: [communications]
      summary: List scoped public alerts pending approval for authorized approvers
      responses:
        "200":
          description: Pending public alerts
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/CommunicationMessage" }
        "403": { description: Public alert approval authority required }
  /communications/{id}/public-alert-review:
    post:
      operationId: reviewPublicAlert
      tags: [communications]
      summary: Approve or reject a scoped pending public alert
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PublicAlertReviewInput" }
      responses:
        "200":
          description: Public alert reviewed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CommunicationMessage" }
        "400": { description: Invalid public alert review }
        "403": { description: Public alert approval authority required }
        "404": { description: Public alert not found }
  /radio-traffic:
    get:
      operationId: listRadioTraffic
      tags: [communications]
      summary: List radio traffic in the caller's authorized scope
      parameters:
        - in: query
          name: priority
          required: false
          schema: { type: string, enum: [emergency, urgent, operational, information] }
      responses:
        "200":
          description: Radio traffic
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/RadioTrafficLog" }
        "403": { description: Operational authority required }
    post:
      operationId: createRadioTraffic
      tags: [communications]
      summary: Record authenticated radio traffic
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RadioTrafficLogInput" }
      responses:
        "201":
          description: Radio traffic recorded
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RadioTrafficLog" }
        "400": { description: Invalid input }
        "403": { description: Scope denied }
components:
  schemas:
    CommunicationMessage:
      type: object
      required: [id, clientId, body, priority, channel, deliveryState, audience, countryCode, parishId, senderUserId, senderName, publicStatus, expiresAt, approvedAt, approvedByName, rejectedAt, rejectedByName, rejectionReason, createdAt, updatedAt, acknowledgementCount, acknowledgedByCurrentUser]
      properties:
        id: { type: string }
        clientId: { type: string }
        body: { type: string }
        priority: { type: string, enum: [emergency, urgent, operational, information] }
        channel: { type: string, enum: [app, radio, sms] }
        deliveryState: { type: string, enum: [recorded, requested, delivered] }
        audience: { $ref: "#/components/schemas/CommunicationAudience" }
        countryCode: { type: string }
        parishId: { type: ["string", "null"] }
        senderUserId: { type: string }
        senderName: { type: string }
        publicStatus: { type: ["string", "null"], enum: [pending_approval, published, rejected, expired, null] }
        expiresAt: { type: ["string", "null"] }
        approvedAt: { type: ["string", "null"] }
        approvedByName: { type: ["string", "null"] }
        rejectedAt: { type: ["string", "null"] }
        rejectedByName: { type: ["string", "null"] }
        rejectionReason: { type: ["string", "null"] }
        createdAt: { type: string }
        updatedAt: { type: string }
        acknowledgementCount: { type: number }
        acknowledgedByCurrentUser: { type: boolean }
    CommunicationMessageInput:
      type: object
      required: [clientId, body, priority, channel, audience, countryCode]
      properties:
        clientId: { type: string, minLength: 1, maxLength: 120 }
        body: { type: string, minLength: 1, maxLength: 4000 }
        priority: { type: string, enum: [emergency, urgent, operational, information] }
        channel: { type: string, enum: [app, radio, sms] }
        audience: { $ref: "#/components/schemas/CommunicationAudience" }
        countryCode: { type: string, minLength: 2, maxLength: 8 }
        parishId: { type: string, maxLength: 120 }
        expiresAt: { type: string }
    CommunicationAudience:
      type: string
      enum: [authorized_scope, incident_commanders, command_team, field_teams, unified_command_partners, public]
    PublicAlertReviewInput:
      type: object
      required: [action]
      properties:
        action: { type: string, enum: [approve, reject] }
        rejectionReason: { type: string, minLength: 1, maxLength: 1000 }
    PublicAlert:
      type: object
      required: [id, body, priority, channel, countryCode, parishId, issuedAt, expiresAt]
      properties:
        id: { type: string }
        body: { type: string }
        priority: { type: string, enum: [emergency, urgent, operational, information] }
        channel: { type: string, enum: [app, radio, sms] }
        countryCode: { type: string }
        parishId: { type: ["string", "null"] }
        issuedAt: { type: string }
        expiresAt: { type: string }
    CommunicationAcknowledgement:
      type: object
      required: [id, messageId, userId, acknowledgedAt]
      properties:
        id: { type: string }
        messageId: { type: string }
        userId: { type: string }
        acknowledgedAt: { type: string }
    RadioTrafficLog:
      type: object
      required: [id, clientId, station, direction, body, occurredAt, priority, countryCode, parishId, recorderUserId, recorderName, createdAt]
      properties:
        id: { type: string }
        clientId: { type: string }
        station: { type: string }
        direction: { type: string, enum: [incoming, outgoing] }
        body: { type: string }
        occurredAt: { type: string }
        priority: { type: string, enum: [emergency, urgent, operational, information] }
        countryCode: { type: string }
        parishId: { type: ["string", "null"] }
        recorderUserId: { type: string }
        recorderName: { type: string }
        createdAt: { type: string }
    RadioTrafficLogInput:
      type: object
      required: [clientId, station, direction, body, occurredAt, priority, countryCode]
      properties:
        clientId: { type: string, minLength: 1, maxLength: 120 }
        station: { type: string, minLength: 1, maxLength: 160 }
        direction: { type: string, enum: [incoming, outgoing] }
        body: { type: string, minLength: 1, maxLength: 4000 }
        occurredAt: { type: string }
        priority: { type: string, enum: [emergency, urgent, operational, information] }
        countryCode: { type: string, minLength: 2, maxLength: 8 }
        parishId: { type: string, maxLength: 120 }
...